:hammer: Rename GlobalCopywriter related classes, add platform prefix for better distinction
